Remove Node dependencies from operator.h
This moves code around so that operator.h/cpp no longer requires a full
definition of Node* nor does it include alias analysis or the pretty printer.

This should make it possible to include in the mobile build.

Functionality for checking if operators match Node and to look up
and operator for a Node have moved to the Node object.

ghstack-source-id: 01e5cfeeef6c1a4c0f120ea07c134cdf905a89cd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32682