Add distributed data parallel benchmark tool
The need for this tool was motivated by #28883. In the past, we have
done ad-hoc benchmarking, but it's time for something more structured.

It would be nice to add more model architectures so that we can get a
full picture of the performance impact of a code change simply by
running this suite a few times.

ghstack-source-id: 1487ab903587da2d55781d781e0d5a79f464e21e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30660